refactor(dashboard-rework): Moved the category query down
