+++ apps/website/src/app/accounts/[tag]/dashboard/(overview)/page.tsx
@@ -3,12 +3,10 @@ import {
   getTotalChests,
   getPeakOpeningHourData,
 } from "@/actions/chest";
-import ChestCountCategoryChart from "@/components/chest-count-category-chart";
 import { getChestCountPerEvent, getPossibleChestCount } from "@/actions/event";
 import { getAccountByTag } from "@/actions/account";
 import LatestChestCard from "@/components/latest-chest-card";
 import TotalChestCard from "@/components/total-chest-card";
-import { getChestCountPerCategory } from "@/actions/category";
 import { getChestCountPerReward } from "@/actions/reward";
 import ChestCountEventTable from "@/components/chest-count-event-table";
 import ChestCountRewardChart from "@/components/chest-count-reward-chart";
@@ -16,6 +14,7 @@ import PeakOpeningHourCard from "@/components/peak-opening-hour-card";
 import AccountInformationItem from "@/components/account-information-item";
 import { FilterConfig } from "@/types/common";
 import RarityCard from "@/components/rarity-card";
+import CategoryCard from "@/components/category-card";
 
 export default async function Page({
   params,
@@ -33,7 +32,6 @@ export default async function Page({
   const chest = await getLatestChest({ accountId: account.id });
   const possibleChestCount = await getPossibleChestCount();
   const chestCount = await getTotalChests({ accountId: account.id });
-  const categories = await getChestCountPerCategory({ accountId: account.id });
   const events = await getChestCountPerEvent({ accountId: account.id });
   const rewards = await getChestCountPerReward({ accountId: account.id });
   const peakOpeningHourData = await getPeakOpeningHourData({
@@ -56,7 +54,7 @@ export default async function Page({
       </div>
       <div className="grid grid-cols-1 lg:grid-cols-2 gap-4">
         <RarityCard filters={filters} />
-        <ChestCountCategoryChart categories={categories} />
+        <CategoryCard filters={filters} />
       </div>
       <ChestCountRewardChart rewards={rewards} />
       <ChestCountEventTable events={events} accountCount={1} />
