+++ apps/website/src/components/category-card.tsx
@@ -1,0 +1,36 @@
+import { getChestCountPerCategory } from "@/actions/category";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import { FilterConfig } from "@/types/common";
+import ChestCountCategoryChart from "./category-chart";
+import { Suspense } from "react";
+import { Skeleton } from "./ui/skeleton";
+
+export interface CategoryCardProps {
+  filters: FilterConfig;
+}
+
+export default async function CategoryCard({ filters }: CategoryCardProps) {
+  const categories = await getChestCountPerCategory(filters);
+
+  return (
+    <Card className="shadow-md">
+      <CardHeader>
+        <CardTitle>Categories</CardTitle>
+        <CardDescription>
+          Shows the number of treasure chests opened per category
+        </CardDescription>
+      </CardHeader>
+      <CardContent>
+        <Suspense fallback={<Skeleton className="aspect-video" />}>
+          <ChestCountCategoryChart categories={categories} />
+        </Suspense>
+      </CardContent>
+    </Card>
+  );
+}
